moved integration tests into their own source folder
