--- src/test/java/org/daemio/merch/MerchServiceApplicationTests.java
@@ -1,12 +1,0 @@
-package org.daemio.merch;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class MerchServiceApplicationTests {
-
-	@Test
-	void contextLoads() {}
-
-}
